Added initial files
So far I have a basic header, as well as a div that holds a paragraph element and a text form box with a submit button. This will later link to the page that shows scraped data of info/facts that are not outdated since the user graduated high school.